requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
ccxt==3.0.36
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
numpy==1.24.2
pandas==2.0.0
parsimonious==0.9.0
protobuf==4.22.1
py-solc-x==1.1.1
pycares==4.3.0
pycparser==2.21
pycryptodome==3.17
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
semantic-version==2.10.0
six==1.16.0
toolz==0.12.0
tzdata==2023.3
urllib3==1.26.15
web3==6.2.0
websockets==10.4
yarl==1.8.2
